\name{structSIR}
\docType{data}
\alias{structSIR}
\title{Forward simulation of SIR dynamics in a structured population}
\description{
This is a phylogenetic tree loaded as an object of class "phylo".
It was generated by the forward-simulation program MASTER v6.0.0 (Vaughan and Drummond 2013) under a susceptible-infected-removed (SIR) model of two subpopulations with divergent transmission rates.
Tip labels of the tree contain the substring \code{I_1} if the lineage was sampled from the high-transmission rate subpopulation and \code{I_0} otherwise.
Thus, the \code{I_1} tips should be assigned to clusters and the \code{I_0} tips should be assigned to the background.
}
\references{
Vaughan TG, Drummond AJ. \emph{A stochastic simulator of birth–death master equations with application to phylodynamics.} Molecular biology and evolution. 2013 Mar 16;30(6):1480-93.

McCloskey RM, Poon AF. \emph{A model-based clustering method to detect infectious disease transmission outbreaks from sequence variation.} PLoS Computational Biology. 2017 Nov 13;13(11):e1005868.
}
\keyword{datasets}
